pyproject.toml
[tool.poetry]
name = "nonebot-plugin-chatpdf"
version = "1.0.0"
description = "A nonebot plugin for chatpdf"
authors = ["Alpaca <alpaca@bupt.edu.cn>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0rc3"
nonebot-adapter-onebot = "^2.2.1"
openai = "^1.30.1"
numpy = "^1.24.2"
faiss-cpu = "^1.7.3"
pymupdf = "^1.21.1"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true 


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
